fix: togglelike api function

+++ api/like/likeController.js
@@ -1,3 +1,16 @@
+import {notNativeNum} from '../function/validCheck.js';
+
+async function isLikeExist(db, userId, postId)
+{
+    const query = `SELECT COUNT(*) FROM likes_mapping WHERE userId = ${userId} AND postId = ${postId}`;
+    const [result] = await db.query(query);
+
+    if(result[0]['COUNT(*)'] === 0)
+        return false;
+    else
+        return true;
+}
+
 export async function toggleLike(req, res)
 {
     const postId = parseInt(req.params.postId,10);
@@ -9,7 +22,6 @@ export async function toggleLike(req, res)
             message : "invalid_postId",
             data : null
         });
-        return false
     }
 
     // 해당 postId를 갖는 post가 존재하는지 확인.
@@ -21,37 +33,48 @@ export async function toggleLike(req, res)
             message : "not_found_post",
             data : null
         });
-        return false;
     }
 
     // likes_mapping 테이블에 해당 postId와 userId를 컬럼값으로 들고 있는 데이터가 있는 경우 삭제, 그렇지 않으면 등록.
-    query = `
-    START TRANSACTION;
+    // -> 상태에 따라 delete or insert로 수정.
 
-    DELETE FROM likes_mapping 
-    WHERE userId = ${userId} AND postId = ${postId};
+    if(await isLikeExist(req.db,userId, postId))
+    {
+        query = `
+        DELETE FROM likes_mapping 
+        WHERE userId = ${userId} AND postId = ${postId};
 
-    UPDATE posts
-    SET likes_count = likes_count - 1
-    WHERE id = ${postId}
-    AND EXISTS (
-        SELECT 1 FROM likes_mapping WHERE userId = ${userId} AND postId = ${postId}
-    );
+        UPDATE posts
+        SET \`like\` = \`like\` - 1
+        WHERE id = ${postId}
+        `
+    }
+    else
+    {
+        query = `
+        INSERT INTO likes_mapping (userId, postId)
+        SELECT ${userId}, ${postId}
+        WHERE NOT EXISTS (
+            SELECT 1 FROM likes_mapping WHERE userId = ${userId} AND postId = ${postId}
+        );
 
-    INSERT INTO likes_mapping (userId, postId)
-    SELECT ${userId}, ${postId}
-    WHERE NOT EXISTS (
-        SELECT 1 FROM likes_mapping WHERE userId = ${userId} AND postId = ${postId}
-    );
+        UPDATE posts
+        SET \`like\` = \`like\` + 1
+        WHERE id = ${postId}
+        `
+    }
+    try{
+        await req.db.query(query);
+        res.status(200).json({
+            message : "toggle_like_success",
+            data : null
+        });
+    }catch(error){
+        console.error(error);
+        res.status(500).json({
+            message : "Internal Server Error",
+            data : null
+        });
+    }
 
-    UPDATE posts
-    SET likes_count = likes_count + 1
-    WHERE id = ${postId}
-    AND NOT EXISTS (
-        SELECT 1 FROM likes_mapping WHERE userId = ${userId} AND postId = ${postId}
-    );
-
-    COMMIT;`
-
-    await req.db.query(query);
 }
